trace loadBookingData
We want to gather more visibility into our requests. We leverage
app.tracer and the request context stored in r.Context() to
create child spans for the loadBookingData() function.

Note that thanks to otelmux the context will contain a parent
span that otel can use to correctly link in a trace chain.

One requirement for this to work is to propagate the request
Context down the call chain.