trace booking render templates
We spend some time each request rendering templates. Exactly how much?

Split templates rendering in a dedicated function and add tracing to it.
To create spans we need to leverage app.tracer. To link this span with
its parent we need to pass the request context, to be used during span
creation.