provide URL information for getAPIContent
Now that we are tracing getAPIContent we notice is heavily
used during a request. But one key information we are lacking
is which endpoint this request is contacting.
This could help spot performance issues in dependencies or
help our troubleshooting efforts.

To provide this information we add an attribute to the
span within the function.