website: instrument website (#1)
* use go 1.22.4

* add otel setup code

See https://opentelemetry.io/docs/languages/go/getting-started/

* add measures setup

We need to init metrics before usage. createMeasures()
init all relevant metrics and return a pointer to a struct
containing all metrics for further use in the app.

* website: count requests to endpoint

* go mod tidy

* refactor main

Split main logic in a separate run function. If os.Exit is called, deferred
functions are not executed. We need deferred functions to properly shut down
otel collection.

* hook otel monitoring

Run otel SDK setup and defer executing accumulated shutdown functions.
If an error happens during shutdown is not recoverable, so log the
error and proceed.

Add tracer and measures to app for further use.

* trace loadBookingData

We want to gather more visibility into our requests. We leverage
app.tracer and the request context stored in r.Context() to
create child spans for the loadBookingData() function.

Note that thanks to otelmux the context will contain a parent
span that otel can use to correctly link in a trace chain.

One requirement for this to work is to propagate the request
Context down the call chain.

* trace booking render templates

We spend some time each request rendering templates. Exactly how much?

Split templates rendering in a dedicated function and add tracing to it.
To create spans we need to leverage app.tracer. To link this span with
its parent we need to pass the request context, to be used during span
creation.

* trace getAPIContent

The last code where we spend time during a request is getAPIContent.

Again we are passing here the request context to retrieve the
request trace and add traces as children of it.

This change is more broad as we need to update all call sites to
pass the context.

* provide URL information for getAPIContent

Now that we are tracing getAPIContent we notice is heavily
used during a request. But one key information we are lacking
is which endpoint this request is contacting.
This could help spot performance issues in dependencies or
help our troubleshooting efforts.

To provide this information we add an attribute to the
span within the function.

* remove infoLog, use slog

* remove errorLog, use slog

* otel ship logs

we must use the HTTP exporter as the gRPC one is still WIP

* cleanup comments & debug logs

* go mod tidy

* no need to set default slog

* propagate traces on HTTP

* deps

* chore: go mod tidy

* chore: add indent

---------

Co-authored-by: Claudio Beatrice <claudi0.beatric3@gmail.com>